% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mooring.R
\name{check_mooring}
\alias{check_mooring}
\title{Identify whether mooring qualifiers are good or bad}
\usage{
check_mooring(quals)
}
\arguments{
\item{quals}{a vector with numeric mooring qualifier codes}
}
\value{
a logical vector indicating good (\code{TRUE}) or bad (\code{FALSE}) qualifier codes
}
\description{
King County mooring data relies on 3 digit qualifier codes to indicate
whether data is good or bad. The first digit indicates overall quality:
1 = good, 2 = probably good, 3 = suspect, 4 = bad, 9 = missing.
\code{check_mooring} takes a vector of qualifiers as an input, and
returns a vector of the same length with values that indicate whether
each qualifier is good (\code{TRUE}) or bad (\code{FALSE}). For the
purposes of this function, 1 and 2 are good and the rest are bad. This
is particularly useful for filtering out bad data in tibbles. For more
information about qualifier codes, see the \href{https://bit.ly/3bi9Xaz}{mooring website}.
}
\examples{

quals <- c(210, 210, 210, 400, 333)
is_good <- check_mooring(quals)
}
