% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_parms-data.R
\docType{data}
\name{discrete_parms}
\alias{discrete_parms}
\title{Data - discrete parameters in the Monitoring Portal}
\format{
A tibble with 3 variables and 57 rows
}
\usage{
data(discrete_parms)
}
\description{
Tibble containing the parameter ID numbers (ParmID), short parameter names
(ParmName), and long parameter names (DisplayName) for every parameter
available in the Monitoring Portal. The function
\code{download_discrete()} uses a vector of ParmName to determine
what parameters to download from the Monitoring Portal.
}
\examples{

data(discrete_parms)
nitrogen_only <- discrete_parms[grepl("Nitrogen", discrete_parms$DisplayName),]
parms_to_download <- nitrogen_only$ParmName
}
\keyword{datasets}
