% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_discrete.R
\name{download_discrete}
\alias{download_discrete}
\title{Download discrete marine data from the Monitoring Portal}
\usage{
download_discrete(sites, parms_in, fname)
}
\arguments{
\item{sites}{a character vector with locator codes for stations. These
stations are the ones for which data will be downloaded.}

\item{parms_in}{a character vector with short parameter names (ParmName)
for which you want data. See \code{discrete_parms}, which is data built
into this package, for a complete list.}

\item{fname}{a string indicating the filename where the data will be saved.
The filename is relative to the current working directory,
\code{getwd()}.}
}
\description{
King County VPN connection required! This function allows you to easily
download any subset of parameters for a given site (or sites) from the
Marine Monitoring Portal. The data will be saved as a .csv file so you
don't have to keep re-downloading it. Speedy and convenient!
Automatically removes bad and missing data with qualifier codes 4 or 9.
Based on code by Curtis DeGasperi.
}
\examples{

sites <- c("NSEX01", "JSUR01", "LSNT01")
parms_in <- c("NNN", "DO")
fname <- "my_discrete_data.csv"

\dontrun{download_discrete(sites, parms_in, fname)}
}
