% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_CTD_composite.R
\name{import_CTD_composite}
\alias{import_CTD_composite}
\title{Import CTD composite data from file}
\usage{
import_CTD_composite(fname)
}
\arguments{
\item{fname}{The filename to import, as a string}
}
\value{
A tibble with imported data and clean column names
}
\description{
Imports CTD data composite from a tab-delimited .txt file or
a .csv file. The format imported by this script is the binned
data composite stored in the shared drive, either the "\emph{_all"
or "}_monthly" files with any bin width. Currently this data
is only available in the internal CTD data repository. These
data do not include any "bad" or "questionable" data.
}
\examples{

file <- system.file("extdata", "test_CTD_composite_data.csv", package = "kcmarine")
data <- import_CTD_composite(file)
}
