% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_discrete.R
\name{import_discrete}
\alias{import_discrete}
\title{Import discrete bottle data from file}
\usage{
import_discrete(fname)
}
\arguments{
\item{fname}{The filename to import, as a string}
}
\value{
A tibble with imported data and clean column names
}
\description{
Imports discrete bottle data from a .csv file. The file format is expected
to be the one produced by \code{download_discrete()}. If you're having
issues loading data, it may be because you have edited the file in
Excel. Try downloading a clean .csv file and not opening it in Excel
prior to loading in R.
}
\examples{

file <- system.file("extdata", "test_discrete_data.csv", package = "kcmarine")
data <- import_discrete(file)
}
