% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mooring_PW.R
\name{import_mooring_PW}
\alias{import_mooring_PW}
\title{Import Point Williams mooring data from file}
\usage{
import_mooring_PW(fname)
}
\arguments{
\item{fname}{The filename to import, as a string}
}
\value{
A tibble with imported data and clean column names
}
\description{
Imports Point mooring data from a .txt file.
Data is expected to contain ALL the parameters
available for download from the public mooring website
(https://green2.kingcounty.gov/marine-buoy/). Columns and header must
be the same as the original data download. When in doubt, download a
new file! Again, make sure you download ALL the parameters. The
function cleans up the column names for easy use in R, and adds a
simple date column.
}
\details{
\preformatted{If you're having trouble and receiving an error about the date format,
try re-downloading the data and NOT editing it in Excel prior to
loading in R. Excel messes up the datetime formatting.
}
}
\examples{

file <- system.file("extdata", "test_mooring_PW_data.txt", package = "kcmarine")
data <- import_mooring_PW(file)
}
